Expanded Cheat Menu:
### Added:
- Help>Interface>GUI Dock Side: Change which side (most) GUI menus are on.

### Fixed:
- Edit ECM Settings blanked out UserSettings when using HelperMod (just in-game, the actual settings are fine after restart).

Lib:
- Changed my msg func to use an object that works in the main menu as well as in-game.

Move HUD Buttons v0.1:
This moves the middle to the left, and the left to the middle.
Pause On Load v0.1:
Pauses the game when you start a new game or load a saved one.
Rightside Menus v0.1:
Not everyone likes change for change's sake. Designers gotta design?
I'd like to know the reasoning behind moving the menus from the right side to the left side...

ExpandedCheatMenu@choggi.org