Expanded Cheat Menu:
### Added:
- Game>Camera>Toggle Map Edge Limit: Removes pushback limit at the edge of the map.
- Debug>View All Entities: Loads a blank map and places all entities in it.

### Changed:
- RC Transport Storage Capacity renamed to RC Storage Capacity (RC Constructors).

### Removed:
- RC Rover Drone Recharge Free: They don't use batteries anymore...

Remove Unwanted Colonists v0.2:
The pod blinked out of existence instead of flying away.
The rocket fx wasn't turned on.
Widened the radius it'll use to hover around waiting.
Fixed issue if pod arrived and victim was in a building.

RC Constructor Routes v0.1:
For some reason it intentionally isn't supposed to be able to create routes...

ExpandedCheatMenu@choggi.org