Expanded Cheat Menu:
- Code cleanup.



Mods:
- A couple random console requestes.

Store Building States:
Adds menus to buildings that will save/remove the state of that building (power,workshift,priority) in a profile.
You can switch profiles using the HUD button, and there's no limit on the amount of profiles or buildings in a profile.
Includes checkbox to add all buildings of the same type using the same state as the selected building.

ExpandedCheatMenu@choggi.org